New tables, split down PublicUser Module.php into traits
while adding new schema and migration
